move implementation of host barrier to a .cpp file
