+++ cpp/include/utilities/host_barrier.hpp
@@ -18,92 +18,12 @@
 #include <raft/handle.hpp>
 #include <rmm/cuda_stream_view.hpp>
 
-#include <vector>
-
 namespace cugraph {
 namespace experimental {
 
 // FIXME: a temporary hack till UCC is integrated into RAFT (so we can use UCC barrier for DASK and
 // MPI barrier for MPI)
-void host_barrier(raft::comms::comms_t const& comm, rmm::cuda_stream_view stream_view)
-{
-  stream_view.synchronize();
-
-  auto const comm_size = comm.get_size();
-  auto const comm_rank = comm.get_rank();
-
-  // k-tree barrier
-
-  int constexpr k = 2;
-  static_assert(k >= 2);
-  std::vector<raft::comms::request_t> requests(k - 1);
-  std::vector<std::byte> dummies(k - 1);
-
-  // up
-
-  int mod = 1;
-  while (mod < comm_size) {
-    if (comm_rank % mod == 0) {
-      auto level_rank = comm_rank / mod;
-      if (level_rank % k == 0) {
-        auto num_irecvs = 0;
-        ;
-        for (int i = 1; i < k; ++i) {
-          auto src_rank = (level_rank + i) * mod;
-          if (src_rank < comm_size) {
-            comm.irecv(dummies.data() + (i - 1),
-                       sizeof(std::byte),
-                       src_rank,
-                       int{0} /* tag */,
-                       requests.data() + (i - 1));
-            ++num_irecvs;
-          }
-        }
-        comm.waitall(num_irecvs, requests.data());
-      } else {
-        comm.isend(dummies.data(),
-                   sizeof(std::byte),
-                   (level_rank - (level_rank % k)) * mod,
-                   int{0} /* tag */,
-                   requests.data());
-        comm.waitall(1, requests.data());
-      }
-    }
-    mod *= k;
-  }
-
-  // down
-
-  mod /= k;
-  while (mod >= 1) {
-    if (comm_rank % mod == 0) {
-      auto level_rank = comm_rank / mod;
-      if (level_rank % k == 0) {
-        auto num_isends = 0;
-        for (int i = 1; i < k; ++i) {
-          auto dst_rank = (level_rank + i) * mod;
-          if (dst_rank < comm_size) {
-            comm.isend(dummies.data() + (i - 1),
-                       sizeof(std::byte),
-                       dst_rank,
-                       int{0} /* tag */,
-                       requests.data() + (i - 1));
-            ++num_isends;
-          }
-        }
-        comm.waitall(num_isends, requests.data());
-      } else {
-        comm.irecv(dummies.data(),
-                   sizeof(std::byte),
-                   (level_rank - (level_rank % k)) * mod,
-                   int{0} /* tag */,
-                   requests.data());
-        comm.waitall(1, requests.data());
-      }
-    }
-    mod /= k;
-  }
-}
+void host_barrier(raft::comms::comms_t const& comm, rmm::cuda_stream_view stream_view);
 
 }  // namespace experimental
 }  // namespace cugraph
